+++ src/components/LoginForm/LoginForm.css
@@ -1,17 +1,64 @@
-.rowLogin{
-    display: flex;
-    align-items: center;
-    text-align: left; 
-   
-    
+.loginForm__welcome{
+    font-family: var(--primary-font);
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 24px;
+    letter-spacing: 0em;
+    text-align: left;
+    margin: 0;
 }
 
-.col_img{
 
-    justify-content: right;
+.loginForm__title{
+    font-family: var(--primary-font);
+    font-size: 32px;
+    font-weight: 500;
+    line-height: 48px;
+    letter-spacing: 0em;
+    text-align: left;
+    margin: 0;
 }
+
+.fieldLabel {
+    display: none;
+}
+
+.loginButton{
+    max-width: 448px !important;
+    width: 100% !important;
+    border-radius: 8px !important;
+    margin: 10px 0 !important;
+    font-size: 24px;
+    font-weight: 600;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+
+.registerText{
+    font-family: var(--secondary-font-2);
+    font-size: 24px;
+    line-height: 28px;
+    color: #616161;
+}
+
+.registerText p{
+    max-width: 450px;
+    margin: 0 auto;
+}
+
+.registerText a {
+    text-decoration: none;
+    color: var(--primary-color);
+}
+
+.fieldControl > input{
+    width: 100% !important;
+    height: 56px !important;
+    margin-bottom: 10px;
+}
+
 .img_manos{
-    height: 100vh;
-    width: 120%;
-    
+    width: 100%;
+    max-width: 708px;
+    height: 820px;
+    object-fit: cover;
 }
