+++ src/index.css
@@ -1,5 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Mulish:wght@400;600;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
 
 body {
   margin: 0;
@@ -18,6 +19,7 @@ code {
 :root{
   --primary-font: 'Poppins', sans-serif; 
   --secondary-font: 'Mulish', sans-serif;
+  --secondary-font-2: 'Roboto', sans-serif;
   --text-1: #000; 
   --primary-color: #FF0000;
   --background-color: #FFF;
